Feat/sharing

+++ src/models/Images.ts
@@ -39,10 +39,7 @@ export interface IImages extends mongoose.Document {
 export const deleteExistingImages = async (parentUiId: string): Promise<void> => {
     const imagesToDelete = await getImagesByParentUiId(parentUiId);
 
-    const deletion: Array<Promise<void>> = [];
-    imagesToDelete.forEach((imageToDelete) => {
-        deletion.concat(deleteImage(imageToDelete));
-    });
+    const deletion = imagesToDelete.map((imageToDelete) => deleteImage(imageToDelete));
 
     await Promise.all(deletion);
 };
@@ -58,8 +55,12 @@ export const getImageByUiId = async (uiId: string): Promise<IImages> => {
 
 export const deleteImage = async (image: IImages): Promise<void> => {
     try {
-        fs.unlinkSync(image.path);
-        fs.unlinkSync(image.thumbnailPath);
+        if (fs.existsSync(image.path)) {
+            fs.unlinkSync(image.path);
+        }
+        if (fs.existsSync(image.thumbnailPath)) {
+            fs.unlinkSync(image.thumbnailPath);
+        }
 
         await image.remove();
     } catch (err) {
